Fix io errors and change syntax

+++ src/translate/fragments/write_int.c
@@ -2,7 +2,7 @@ void write_int(int64_t v) {
     char buf[32];
     int off = 32;
     int sign = v < 0;
-    uint32_t work = v;
+    uint64_t work = v;
     if (sign) work = -work;
     do {
         buf[--off] = work % 10 + '0';
